+++ config/routes.rb
@@ -89,7 +89,7 @@ Rails.application.routes.draw do
 
     resources :claims, only: [:index, :show] do
       resources :checks, only: [:index, :show, :create], param: :check, constraints: {check: %r{#{Admin::ChecksController::CHECKS_SEQUENCE.join("|")}}}
-      resources :decisions, only: [:create]
+      resources :decisions, only: [:create, :new]
       get "search", on: :collection
     end