+++ spec/features/admin_claim_check_spec.rb
@@ -22,7 +22,7 @@ RSpec.feature "Admin checks a claim" do
         find("a[href='#{admin_claim_path(claim_to_approve)}']").click
         choose "Approve"
         fill_in "Decision notes", with: "Everything matches"
-        perform_enqueued_jobs { click_on "Submit" }
+        perform_enqueued_jobs { click_on "Confirm decision" }
 
         expect(claim_to_approve.decision.created_by).to eq(user)
         expect(claim_to_approve.decision.notes).to eq("Everything matches")
@@ -53,7 +53,7 @@ RSpec.feature "Admin checks a claim" do
       find("a[href='#{admin_claim_path(claim_to_reject)}']").click
       choose "Reject"
       fill_in "Decision notes", with: "TRN doesn't exist"
-      perform_enqueued_jobs { click_on "Submit" }
+      perform_enqueued_jobs { click_on "Confirm decision" }
 
       expect(claim_to_reject.decision.created_by).to eq(user)
       expect(claim_to_reject.decision.notes).to eq("TRN doesn't exist")
@@ -68,7 +68,7 @@ RSpec.feature "Admin checks a claim" do
       expect(mail.body.raw_source).to match("not been able to approve")
     end
 
-    scenario "User can see checks for a claim" do
+    scenario "User can complete the claim checks and approve a claim" do
       claim = create(:claim, :submitted, policy: MathsAndPhysics)
 
       visit admin_claims_path
@@ -91,6 +91,14 @@ RSpec.feature "Admin checks a claim" do
       click_on "Complete employment check and continue"
 
       expect(page).to have_content("Claim decision")
+
+      choose "Approve"
+      fill_in "Decision notes", with: "All checks passed!"
+      click_on "Confirm decision"
+
+      expect(page).to have_content("Claim has been approved successfully")
+      expect(claim.decision).to be_approved
+      expect(claim.decision.created_by).to eq(user)
     end
 
     scenario "User can see completed checks" do
@@ -113,7 +121,7 @@ RSpec.feature "Admin checks a claim" do
       claim_with_decision = create(:claim, :submitted, decision: build(:decision, result: :approved, notes: "Everything matches"))
       visit admin_claim_path(claim_with_decision)
 
-      expect(page).not_to have_button("Submit")
+      expect(page).not_to have_button("Confirm decision")
       expect(page).to have_content("Claim decision")
       expect(page).to have_content("Approved")
       expect(page).to have_content(claim_with_decision.decision.notes)
@@ -193,7 +201,7 @@ RSpec.feature "Admin checks a claim" do
 
         choose "Approve"
         fill_in "Decision notes", with: "Identity confirmed via phone call"
-        click_on "Submit"
+        click_on "Confirm decision"
 
         expect(claim_without_identity_confirmation.decision.created_by).to eq(user)
         expect(claim_without_identity_confirmation.decision.notes).to eq("Identity confirmed via phone call")
@@ -223,6 +231,66 @@ RSpec.feature "Admin checks a claim" do
         student_loan_claims.each do |c|
           expect(page).to_not have_content(c.reference)
         end
+      end
+    end
+
+    context "when the service operator completes the last check" do
+      context "and the payroll gender is missing" do
+        let!(:claim) { create(:claim, :submitted, payroll_gender: :dont_know) }
+
+        scenario "User is informed that the claim cannot be approved" do
+          perform_last_check
+
+          expect(page).to have_field("Approve", disabled: true)
+          expect(page).to have_content(I18n.t("admin.unknown_payroll_gender_preventing_approval_message"))
+        end
+      end
+
+      context "and the claimant has another approved claim in the same payroll window, with inconsistent personal details" do
+        let(:personal_details) do
+          {
+            national_insurance_number: generate(:national_insurance_number),
+            teacher_reference_number: generate(:teacher_reference_number),
+            date_of_birth: 30.years.ago.to_date,
+            student_loan_plan: StudentLoan::PLAN_1,
+            email_address: "email@example.com",
+            bank_sort_code: "112233",
+            bank_account_number: "95928482",
+            building_society_roll_number: nil
+          }
+        end
+        let!(:approved_claim) { create(:claim, :approved, personal_details.merge(bank_sort_code: "112233", bank_account_number: "29482823")) }
+        let!(:claim) { create(:claim, :submitted, personal_details.merge(bank_sort_code: "582939", bank_account_number: "74727752")) }
+
+        scenario "User is informed that the claim cannot be approved" do
+          perform_last_check
+
+          expect(page).to have_field("Approve", disabled: true)
+          expect(page).to have_content("This claim cannot currently be approved because we’re already paying another claim (#{approved_claim.reference}) to this claimant in this payroll month using different payment details. Please speak to a Grade 7.")
+        end
+      end
+
+      context "and the claimant has not completed GOV.UK Verify" do
+        let!(:claim) { create(:claim, :unverified) }
+
+        scenario "the service operator is told the identity hasn't been confirmed and can approve the claim" do
+          perform_last_check
+
+          expect(page).to have_content("The claimant did not complete GOV.UK Verify")
+          expect(page).to have_content(claim.school.phone_number)
+
+          choose "Approve"
+          fill_in "Decision notes", with: "Identity confirmed via phone call"
+          click_on "Confirm decision"
+
+          expect(claim.decision.created_by).to eq(user)
+          expect(claim.decision.notes).to eq("Identity confirmed via phone call")
+        end
+      end
+
+      def perform_last_check
+        visit admin_claim_check_path(claim, check: Admin::ChecksController::CHECKS_SEQUENCE.last)
+        find("input[type='submit']").click
       end
     end
   end